Move GetHHID to models
