+++ application/models/models_test.go
@@ -95,3 +95,32 @@ func (ms *ModelSuite) EqualNullUUID(expected nulls.UUID, actual *uuid.UUID, msgA
 		ms.Equal(expected, *actual, msgAndArgs...)
 	}
 }
+
+func (ms *ModelSuite) TestGetHHID() {
+	if domain.Env.HouseholdIDLookupURL == "" {
+		ms.T().Skip("skipping test because no HOUSEHOLD_ID_LOOKUP_URL was provided")
+	}
+
+	tests := []struct {
+		name    string
+		staffID string
+		want    string
+	}{
+		{
+			name:    "good",
+			staffID: "32329",
+			want:    "232329",
+		},
+		{
+			name:    "not found",
+			staffID: "9999999",
+			want:    "",
+		},
+	}
+	for _, tt := range tests {
+		ms.T().Run(tt.name, func(t *testing.T) {
+			got := GetHHID(tt.staffID)
+			ms.Equal(tt.want, got)
+		})
+	}
+}
